feat(Lotto): 로또 번호가 중복인지 판단하는 기능 구현
